refactor: :art: moves AgreementTable to Agreements domain directory
