+++ frontend/src/components/Agreements/AgreementTableRow/AgreementTableRow.jsx
@@ -6,10 +6,10 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faChevronDown, faChevronUp, faPen, faTrash } from "@fortawesome/free-solid-svg-icons";
 import { faClock } from "@fortawesome/free-regular-svg-icons";
 import { convertCodeForDisplay, formatDate } from "../../../helpers/utils";
-import TableTag from "../../../components/UI/TableTag";
+import TableTag from "../../UI/TableTag";
 import { useDeleteAgreementMutation } from "../../../api/opsAPI";
 import icons from "../../../uswds/img/sprite.svg";
-import ConfirmationModal from "../../../components/UI/Modals/ConfirmationModal";
+import ConfirmationModal from "../../UI/Modals/ConfirmationModal";
 import useGetUserFullNameFromId from "../../../helpers/user-hooks";
 import { useIsUserAllowedToEditAgreement } from "../../../helpers/agreement-hooks";
 import { DISABLED_ICON_CLASSES } from "../../../constants";
