+++ frontend/src/components/Agreements/AgreementTableRow/index.js
@@ -1,0 +1,2 @@
+export * from "./AgreementTableRow";
+export { default } from "./AgreementTableRow";
