+++ frontend/src/components/Agreements/AgreementsTable/index.js
@@ -1,0 +1,2 @@
+export * from "./AgreementsTable";
+export { default } from "./AgreementsTable";
