+++ frontend/src/pages/agreements/list/AgreementsList.jsx
@@ -6,7 +6,7 @@ import App from "../../../App";
 import { useGetAgreementsQuery } from "../../../api/opsAPI";
 import Breadcrumb from "../../../components/UI/Header/Breadcrumb";
 import sortAgreements from "./utils";
-import AgreementsTable from "./AgreementsTable";
+import AgreementsTable from "../../../components/Agreements/AgreementsTable";
 import AgreementTabs from "./AgreementsTabs";
 import { getCurrentFiscalYear } from "../../../helpers/utils";
 import TablePageLayout from "../../../components/Layouts/TablePageLayout";
